requirements.txt
astroid==3.3.8
async-timeout==5.0.1
blinker==1.9.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
identify==2.6.4
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
numpy==2.2.1
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pylint==3.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.24.0
SQLAlchemy==2.0.36
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.1
Werkzeug==3.1.3
